Moved zone.js to a separate vendor script.
Since Angular requires core-js's reflect scripts for its dependency 
injection, it couldn't be skipped. While Browserify was able to bundle 
core-js and zone.js into the same script, for some reason, rollup 
appears to be struggling with it. Thus, we've switched to the same 
structure as the Angular CLI (which bundles zone.js into a separate 
script).